Reorganize demo app code
