+++ demo/index.html
@@ -1,17 +1,19 @@
 <!--
-    Copyright (c) 2023, RTE (http://www.rte-france.com)
-    This Source Code Form is subject to the terms of the Mozilla Public
-    License, v. 2.0. If a copy of the MPL was not distributed with this
-    file, You can obtain one at http://mozilla.org/MPL/2.0/.
--->
+  ~ Copyright (c) 2023, RTE (https://www.rte-france.com)
+  ~ This Source Code Form is subject to the terms of the Mozilla Public
+  ~ License, v. 2.0. If a copy of the MPL was not distributed with this
+  ~ file, You can obtain one at http://mozilla.org/MPL/2.0/.
+  ~ SPDX-License-Identifier: MPL-2.0
+  -->
+
 <!doctype html>
 <html lang="en">
     <head>
         <meta charset="utf-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <meta name="theme-color" content="#000000" />
-        <meta name="description" content="network map viewer demo app" />
-        <title>SVG Viewers Examples</title>
+        <meta name="description" content="network-viewer demo app" />
+        <title>Network-viewer examples</title>
         <link rel="stylesheet" href="src/style.css" />
         <link rel="stylesheet" href="src/zoom.css" />
         <link rel="stylesheet" href="src/demo.css" />
@@ -20,20 +22,13 @@
         <div class="floating-toc">
             <h3>Contents</h3>
             <ul>
-                <li><a href="#map-viewer">Network Map Viewer</a></li>
-                <li><a href="#nad-viewers">Network Area Diagram Viewers</a></li>
-                <li><a href="#sld-viewers">Single Line Diagram Viewers</a></li>
-                <li><hr /></li>
-                <li><a href="synced-viewers.html">Synced Viewers Demo</a></li>
+                <li><a href="index.html">Network area diagram viewers</a></li>
+                <li><a href="map-viewer.html">Map viewers</a></li>
+                <li><a href="sld.html">Single-line diagram viewers</a></li>
+                <li><a href="synced-viewers.html">Synced viewers demo</a></li>
             </ul>
         </div>
         <div class="flex-container">
-            <div id="map-viewer" class="viewer-section">
-                <h2>Network Map Viewer</h2>
-                <div class="break"></div>
-                <div id="root"></div>
-            </div>
-            <div class="break"></div>
             <div id="nad-viewers" class="viewer-section">
                 <h2>Network Area Diagram Viewers</h2>
                 <div class="break"></div>
@@ -57,7 +52,7 @@
                 <div id="svg-container-nad-multibus-vlnodes-middle-arrow"></div>
                 <div class="break"></div>
                 <div id="svg-container-nad-multibus-vlnodes-limit-percentage"></div>
-                <div id="svg-container-nad-pst-hvdc-multiple-labels"></div>                
+                <div id="svg-container-nad-pst-hvdc-multiple-labels"></div>
                 <div class="break"></div>
                 <div id="svg-container-nad-hoverCallback"></div>
                 <div id="svg-container-nad-pst-hvdc-custom"></div>
@@ -65,16 +60,7 @@
                 <div id="svg-container-nad-double-arrows"></div>
                 <div id="svg-container-nad-components"></div>
             </div>
-            <div class="break"></div>
-            <div id="sld-viewers" class="viewer-section">
-                <h2>Single Line Diagram Viewers</h2>
-                <div class="break"></div>
-                <div id="svg-container-sld"></div>
-                <div id="svg-container-sld-with-callbacks"></div>
-                <div class="break"></div>
-                <div id="svg-container-sldsub-with-callbacks"></div>
-            </div>
         </div>
-        <script type="module" src="/src/index.tsx"></script>
+        <script type="module" src="./src/nad.ts"></script>
     </body>
 </html>
