+++ demo/map-viewer.html
@@ -17,30 +17,6 @@
         <link rel="stylesheet" href="src/style.css" />
         <link rel="stylesheet" href="src/zoom.css" />
         <link rel="stylesheet" href="src/demo.css" />
-        <style>
-            .viewer-section {
-                margin-bottom: 40px;
-            }
-
-            .viewer-section h2 {
-                padding-left: 10px;
-            }
-
-            .synced-container {
-                display: flex;
-                gap: 10px;
-                padding: 10px;
-            }
-
-            .synced-container > div {
-                display: flex;
-                align-items: center;
-            }
-
-            [id^='svg-container-nad-synced-'] {
-                border: 1px solid #ccc;
-            }
-        </style>
     </head>
     <body>
         <div class="floating-toc">
@@ -53,16 +29,14 @@
             </ul>
         </div>
 
-        <div id="synced-viewers" class="viewer-section">
-            <h2>Synced Viewers</h2>
-            <p>Three viewers synchronized with slightly different branch data on the IEEE 14 bus network.</p>
-            <div class="synced-container">
-                <div id="svg-container-nad-synced-1"></div>
-                <div id="svg-container-nad-synced-2"></div>
-                <div id="svg-container-nad-synced-3"></div>
+        <div class="flex-container">
+            <div id="map-viewer" class="viewer-section">
+                <h2>Network map viewer</h2>
+                <div class="break"></div>
+                <div id="root"></div>
             </div>
         </div>
 
-        <script type="module" src="./src/synced-viewers.ts"></script>
+        <script type="module" src="./src/map-viewer.tsx"></script>
     </body>
 </html>
