+++ demo/src/diagram-viewers/data/sld/sld-example_metadata.json
