+++ demo/src/diagram-viewers/data/sld/sld-sub-example_metadata.json
